[CON-8 XTC] Move ICP ledger did from remote/ledger to remote/icp
